requirements.txt
aiohttp==3.6.3
async-cleverbot==0.2.2
async-timeout==3.0.1
asyncpg==0.21.0
discord.py==1.5.0
numpy==1.19.4
opencv-python-headless==4.4.0.46
Pillow==7.2.0
urllib3==1.25.10
pytz==2020.1
python-dateutil==2.8.1
python-dotenv==0.15.0
kaleido==0.1.0
plotly==4.14.1